feat: Populating the file ducks/auth/index.js
