+++ src/store/ducks/auth/actions.js
@@ -1,0 +1,15 @@
+import { AuthTypes } from './types';
+
+export const signIn = (user) => ({
+  type: AuthTypes.SIGNIN,
+  payload: user
+});
+
+export const signUp = (user) => ({
+  type: AuthTypes.SIGNUP,
+  payload: user
+});
+
+export const logOut = () => ({
+  type: AuthTypes.LOGOUT
+});
